delete issue fixed where delete is for only owner user

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -153,9 +153,24 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        // get logged in user form session loggeduser
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        //check if current user id matches with image owner userid
+        //if it matches user can edit image and redirected to image and delete image
+        //else request will be transferred to images/image with error message for user to display
+        if(image.getUser().getId()!=loggedInUser.getId()){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            String deleteError = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", deleteError);
+            return "images/image";
+        }else{
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
